[examples] Fix pipeline issues
