+++ examples/material-next-ts-app-router/app/layout.tsx
@@ -1,17 +1,17 @@
-import { ReactNode } from "react";
-import ThemeProvider from "@/src/utils/ThemeProvider";
+import * as React from 'react';
+import ThemeProvider from '../src/ThemeProvider';
 
 export const metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
-}
+};
 
-export default function RootLayout({ children }: { children: ReactNode }) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
       <body>
         <ThemeProvider>{children}</ThemeProvider>
       </body>
     </html>
-  )
+  );
 }
