+++ examples/material-next-ts-app-router/app/page.tsx
@@ -1,11 +1,10 @@
-import Home from "@/src/views/Home";
+import * as React from 'react';
+import Home from '../src/Home';
 
 export const metadata = {
-    title: 'Home'
-}
+  title: 'Home',
+};
 
 export default function HomePage() {
-    return (
-        <Home/>
-    );
+  return <Home />;
 }
