--- examples/material-next-ts-app-router/src/styles/theme.ts
@@ -1,29 +1,0 @@
-import { Roboto } from 'next/font/google';
-import { red } from '@mui/material/colors';
-import { createTheme } from '@mui/material/styles';
-
-const roboto = Roboto({
-    weight: ['300', '400', '500', '700'],
-    subsets: ['latin'],
-    display: 'swap',
-    fallback: ['Helvetica', 'Arial', 'sans-serif'],
-});
-
-const theme = createTheme({
-    palette: {
-        primary: {
-            main: '#556cd6',
-        },
-        secondary: {
-            main: '#19857b',
-        },
-        error: {
-            main: red.A400,
-        },
-    },
-    typography: {
-        fontFamily: roboto.style.fontFamily,
-    },
-});
-
-export default theme;
